allocatorimpl: extract ValidateReplicationChanges from kvserver
Move validateReplicationChanges (renamed ValidateReplicationChanges) and
its four helper validators (validateAdditionsPerStore, validateRemovals,
validatePromotionsAndDemotions, validateOneReplicaPerNode) plus the two
internal map types (changesByStoreID, changesByNodeID) from
replica_command.go into a new file under
pkg/kv/kvserver/allocator/allocatorimpl. The single caller in
replica_command.go and the existing test in replica_command_test.go are
updated accordingly.

The motivation is to let the allocator simulator (pkg/kv/kvserver/asim)
apply the same rejection rules as the production replicate queue without
depending on pkg/kv/kvserver — asim currently applies any
planner-emitted (ADD, REMOVE) pair, which masks allocator bugs whose
only symptom is a rejected change. asim already imports allocatorimpl
for the allocator itself, so this avoids introducing a new package or a
new import edge into asim.

No behavior change.

Release note: None
Epic: none
